requirements.txt
flask
python-dotenv
shortuuid
anthropic
youtube-transcript-api
moviepy
edge-tts
Pillow
requests
